bizflow/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==26.1
pillow==11.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
